+++ packages/component/src/block/markets/sections/listView.tsx
@@ -1,6 +1,6 @@
 import { ListTile, ListView } from "@/listView";
 import { Divider } from "@/divider";
-import { FC, useCallback, useMemo, useState } from "react";
+import { FC, useCallback, useEffect, useMemo, useState } from "react";
 import { Statistic } from "@/statistic";
 import { SortCondition, SortGroup } from "@/block/markets/sections/sortGroup";
 import { Numeral } from "@/text";
@@ -28,11 +28,29 @@ export const MarketListView: FC<MarketListViewProps & Props> = (props) => {
     return <Divider className="orderly-my-[16px]" />;
   }, []);
 
+  const [innerHeight, setInnerHeight] = useState(window.innerHeight - 180);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setInnerHeight(window.innerHeight - 180);
+    };
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+
   return (
     <>
       <SortGroup onChange={props.onSort} />
       <Divider />
-      <div className="orderly-overflow-y-auto orderly-scrollbar-hidden orderly-hide-scrollbar orderly-flex-grow orderly-my-2">
+      <div className="orderly-overflow-y-auto orderly-overflow-hidden orderly-scrollbar-hidden orderly-hide-scrollbar orderly-h-[calc(100vh-180px)] orderly-my-2"
+      style={{
+        height: innerHeight
+      }}>
         <ListView.separated<API.MarketInfoExt>
           dataSource={props.dataSource}
           renderItem={renderItem}