+++ packages/component/src/block/positions/full/index.tsx
@@ -3,13 +3,16 @@ import { Divider } from "@/divider";
 import { Listview } from "./listview";
 import { FC } from "react";
 import { PositionsViewProps } from "../types";
+import { AssetsProvider } from "@/provider";
 
 export const PositionsViewFull: FC<
   Omit<PositionsViewProps, "onShowAllSymbolChange">
 > = (props) => {
   return (
     <>
-      <Header aggregated={props.aggregated} />
+      <AssetsProvider>
+        <Header aggregated={props.aggregated} />
+      </AssetsProvider>
       <Divider />
       <Listview {...props} />
     </>