+++ packages/component/src/page/trading/fill/sections/datalist/positions/index.tsx
@@ -74,7 +74,7 @@ export const PositionPane = () => {
               toast.error(err.message);
             });
         },
-        contentClassName: "desktop:orderly-w-[364px]",
+        maxWidth: "xs",
       })
       .catch(() => {});
   }, []);